+++ .gitignore
@@ -1,2 +1,3 @@
 /.idea/
 /target/
+/src/test/resources/RandomDataFile.dat
