+++ src/main/java/don/juan/matus/lib/tree/bintree/rotatetree/scapegoat/RandomRotateBinTree.java
@@ -11,13 +11,16 @@ public class RandomRotateBinTree<T extends Comparable<T>> extends ScapegoatTree<
 
     public RandomRotateBinTree() {
         super();
-        alpha = 0.5d;
+        alpha = 0.7d;
         rnd = new Random();
         root = new BinTreeNodeBase<>(null, null, null, null);
     }
 
     protected BinTreeNodeInterface<T> rebalance(BinTreeNodeInterface<T> cursor) {
         boolean isLeft = (cursor == cursor.getParent().getLeft());
+        if (cursor.getParent() == root) {
+            isLeft = true;
+        }
         Long clevel = level;
         cursor = cursor.getParent();
         while (cursor != root) {
