+++ src/test/java/don/juan/matus/lib/tree/bintree/SimpleTest.java
@@ -1,30 +1,18 @@
 package don.juan.matus.lib.tree.bintree;
 
-import don.juan.matus.lib.tree.bintree.mergeabletree.cartesian.CartesianBinTree;
-import don.juan.matus.lib.tree.bintree.mergeabletree.cartesian.RandomMergeBinTree;
-import don.juan.matus.lib.tree.bintree.rotatetree.aa.AATree;
 import don.juan.matus.lib.tree.bintree.rotatetree.avl.AVLBinTree;
 import don.juan.matus.lib.tree.bintree.rotatetree.random.RndBinTree;
 import don.juan.matus.lib.tree.bintree.rotatetree.red_black.RedBlackTree;
-import don.juan.matus.lib.tree.bintree.rotatetree.scapegoat.RandomRotateBinTree;
 import don.juan.matus.lib.tree.bintree.rotatetree.scapegoat.ScapegoatTree;
-import don.juan.matus.lib.tree.bintree.rotatetree.splay.SplayTree;
 import don.juan.matus.lib.tree.bintree.rotatetree.waight.BinTreeW;
 import don.juan.matus.lib.tree.bintree.tst.TreeMapTst;
-import don.juan.matus.lib.tree.bintree.tst.TreeSetTst;
 import junit.framework.TestCase;
 import org.apache.commons.collections.list.TreeList;
 
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.HashMap;
 import java.util.Iterator;
-import java.util.List;
 import java.util.Map;
 import java.util.TreeSet;
 
@@ -1130,476 +1118,6 @@ public class SimpleTest extends TestCase {
         System.out.println("time = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
     }
 
-    public void testBinTree3() {
-        Calendar cBegin = Calendar.getInstance();
-        BinTreeW<Long> btLng = new BinTreeW<Long>();
-        Long rnd;
-        for (int i = 0; i < 10000000; i++) {
-            rnd = Math.round(Math.random() * 10000L);
-            btLng.add(rnd);
-            if (i % 100000 == 0) {
-                System.out.println("i = " + i);
-            }
-        }
-        System.out.println("---------------------------------------------------");
-        System.out.println("maxLevel = " + btLng.getMaxLevel());
-        System.out.println("RotateCount = " + btLng.getRotateCount());
-        System.out.println("---------------------------------------------------");
-        final Iterator<Long> it = btLng.iterator();
-        btLng.passStraight(new BinTreePassEvent<BinTreeNodeInterface<Long>>() {
-            int i = 1;
-
-            public void onPass(BinTreeNodeInterface<Long> theObject) {
-                if (it.hasNext()) {
-                    assertEquals(theObject.getObjectNode(), it.next());
-                }
-                if (i % 100000 == 0) {
-                    System.out.println("i = " + i);
-                }
-                i++;
-            }
-
-            public Long incLevel(BinTreeNodeInterface<Long> theObject) {
-                return null;
-            }
-
-            public Long decLevel(BinTreeNodeInterface<Long> theObject) {
-                return null;
-            }
-
-            public Long incLeft(BinTreeNodeInterface<Long> theObject) {
-                return null;
-            }
-
-            public Long decLeft(BinTreeNodeInterface<Long> theObject) {
-                return null;
-            }
-
-            public Long incRight(BinTreeNodeInterface<Long> theObject) {
-                return null;
-            }
-
-            public Long decRight(BinTreeNodeInterface<Long> theObject) {
-                return null;
-            }
-
-            @Override
-            public void onNodeCompleted(BinTreeNodeInterface<Long> theObject) {
-
-            }
-
-            @Override
-            public void onPassCompleted() {
-
-            }
-        });
-        Calendar cEnd = Calendar.getInstance();
-        System.out.println("---------------------------------------------------");
-        System.out.println("time = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-    }
-
-    private static final int maxRandomDataFile = 10000000;
-
-    public void testCreateRandomDataFile() throws IOException {
-        TreeMapTst<Long, Long> tmRndLng = new TreeMapTst<Long, Long>();
-        List<Long> rList = new ArrayList<Long>();
-        Long rnd;
-        while (tmRndLng.size() < maxRandomDataFile) {
-            rnd = Math.round(Math.random() * maxRandomDataFile);
-            if (tmRndLng.put(rnd, rnd) == null) rList.add(rnd);
-        }
-        tmRndLng = null;
-        System.gc();
-
-        ObjectOutputStream os = new ObjectOutputStream(new FileOutputStream("src/test/resources/RandomDataFile.dat"));
-        os.writeObject(rList);
-        os.flush();
-        os.close();
-
-    }
-
-    private List<Long> readRendomDataFile() throws IOException, ClassNotFoundException {
-        List<Long> result;
-        ObjectInputStream is = new ObjectInputStream(new FileInputStream("src/test/resources/RandomDataFile.dat"));
-        result = (List<Long>) is.readObject();
-        is.close();
-        return result;
-    }
-
-    private <T extends Comparable<T>> void addSeekRemove(String nameTree, List<T> rList, BinTreeBase<T> tree) {
-        Calendar cBegin;
-        Calendar cEnd;
-        T rnd;
-        BinTreeBase.TreeProps tp;
-        int j;
-        Iterator<T> it;
-        //
-        System.out.println(nameTree);
-        cBegin = Calendar.getInstance();
-        for (int i = 0; i < maxRandomDataFile; i++) {
-            rnd = rList.get(i);
-            tree.add(rnd);
-        }
-        cEnd = Calendar.getInstance();
-        System.out.print("time 2.1 (add) = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        tp = tree.treePassage();
-        System.out.println(", maxLevel = " + tp.heght + ", RotateCount = " + tree.getRotateCount());
-        //
-        cBegin = Calendar.getInstance();
-        it = tree.iterator();
-        j = 0;
-        while (it.hasNext()) {
-            it.next();
-            j++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.1 (iterate) = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + j + ") " + tree.getSize());
-        //
-        cBegin = Calendar.getInstance();
-        tree.checkStructure(new BinTreeCheckPassEventTest<>());
-        cEnd = Calendar.getInstance();
-        System.out.println("time 4.1 (check) = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " " + tree.getSize());
-        //
-        tree.setRotateCount(0L);
-        cBegin = Calendar.getInstance();
-        for (int i = 0; i < maxRandomDataFile; i += 100) {
-            for (int l = 0; l < 10; l++) {
-                for (int k = 0; k < 99; k++) {
-                    if (i + k < maxRandomDataFile) {
-                        rnd = rList.get(i + k);
-                        if (!tree.seek(rnd, null)) {
-                            System.out.println(String.format("%s not found!", rnd));
-                        }
-                    }
-                }
-            }
-        }
-        cEnd = Calendar.getInstance();
-        System.out.print("time 5.1 (seek)  = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        tp = tree.treePassage();
-        System.out.println(", maxLevel = " + tp.heght + ", RotateCount = " + tree.getRotateCount());
-        //
-        tree.setRotateCount(0L);
-        cBegin = Calendar.getInstance();
-        it = tree.iterator();
-        j = 0;
-        int r = 0;
-        while (it.hasNext()) {
-            if ((j & 1) == 1) {
-                it.remove();
-                r++;
-            }
-            it.next();
-            j++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.print("time 6.1 (remove) = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + r + ") " + tree.getSize() + ", - " + (r + tree.getSize()));
-        tp = tree.treePassage();
-        System.out.println(", maxLevel = " + tp.heght + ", RotateCount = " + tree.getRotateCount());
-        //
-        cBegin = Calendar.getInstance();
-        tree.checkStructure(new BinTreeCheckPassEventTest<>());
-        cEnd = Calendar.getInstance();
-        System.out.println("time 7.1 (check) = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " " + tree.getSize());
-        //
-        System.gc();
-        System.out.println("<=========================================================================================>");
-
-    }
-
-
-    public void testBinTree7() throws IOException, ClassNotFoundException {
-        Calendar cBegin = Calendar.getInstance();
-        List<Long> rList = readRendomDataFile();
-        Calendar cEnd = Calendar.getInstance();
-        System.out.println("time 1 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        System.out.println("");
-        addSeekRemove("BinTreeBase", rList, new BinTreeBase<>());
-        addSeekRemove("BinTreeW", rList, new BinTreeW<>());
-        addSeekRemove("AVLBinTree", rList, new AVLBinTree<>());
-        addSeekRemove("RedBlackTree", rList, new RedBlackTree<>());
-        addSeekRemove("AATree", rList, new AATree<>());
-//        addSeekRemove("SplayTree", rList, new SplayTree<>());
-        addSeekRemove("ScapegoatTree", rList, new ScapegoatTree<>());
-        addSeekRemove("CartesianBinTree", rList, new CartesianBinTree<>());
-        addSeekRemove("RandomMergeBinTree", rList, new RandomMergeBinTree<>());
-        addSeekRemove("RndBinTree", rList, new RndBinTree<>());
-//        addSeekRemove("RandomRotateBinTree", rList, new RandomRotateBinTree<>());
-        //
-
-/*
-        System.out.println("TreeMap");
-        cBegin = Calendar.getInstance();
-        TreeMapTst<Long, Long> tmLng = new TreeMapTst<Long, Long>();
-        for (int i = 0; i < maxRandomDataFile; i++) {
-            rnd = rList.get(i);
-            tmLng.put(rnd, rnd);
-        }
-        System.out.println("maxLevel = " + tmLng.maxLevel);
-        System.out.println("RotateCount = " + tmLng.rotateCount);
-        cEnd = Calendar.getInstance();
-        System.out.println("time 2.2 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        //
-        cBegin = Calendar.getInstance();
-        it = tmLng.descendingKeySet().iterator();
-        j = 0;
-        while (it.hasNext()) {
-            it.next();
-            j++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.2 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + j + ") " + tmLng.size());
-        tmLng = null;
-        System.gc();
-        System.out.println("<=========================================================================================>");
-        //
- */
-    }
-
-    public void testBinTreeBase() {
-        int j;
-        Iterator<Long> it;
-        TreeMapTst<Long, Long> tmRndLng = new TreeMapTst<Long, Long>();
-        List<Long> rList = new ArrayList<Long>();
-        Long rnd;
-        while (tmRndLng.size() < 10000000) {
-            rnd = Math.round(Math.random() * 10000000L);
-            if (tmRndLng.put(rnd, rnd) == null) rList.add(rnd);
-        }
-        tmRndLng = null;
-        System.gc();
-        //
-        System.out.println("BinTreeBase");
-        Calendar cBegin = Calendar.getInstance();
-        BinTreeBase<Long> baseBLng = new BinTreeBase<Long>();
-        for (int i = 0; i < 10000000; i++) {
-            rnd = rList.get(i);
-            baseBLng.add(rnd);
-        }
-        System.out.println("maxLevel = " + baseBLng.getMaxLevel());
-        Calendar cEnd = Calendar.getInstance();
-        System.out.println("time 2.4 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        //
-        cBegin = Calendar.getInstance();
-        it = baseBLng.iterator();
-        j = 0;
-        while (it.hasNext()) {
-            it.next();
-            j++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.4 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + j + ") " + baseBLng.getSize());
-        System.gc();
-        //
-        cBegin = Calendar.getInstance();
-        it = baseBLng.iterator();
-        j = 0;
-        while (it.hasNext()) {
-            it.next();
-            if (j % 3 == 0) {
-                System.out.println("j=" + j);
-                it.remove();
-            }
-            j++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.5 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + j + ") " + baseBLng.getSize());
-        System.gc();
-        //
-        cBegin = Calendar.getInstance();
-        it = baseBLng.iterator();
-        j = 0;
-        while (it.hasNext()) {
-            it.next();
-            j++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.6 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + j + ") " + baseBLng.getSize());
-        baseBLng = null;
-        System.gc();
-    }
-
-    public void testBinTree77() {
-        Calendar cBegin = Calendar.getInstance();
-        TreeMapTst<Long, Long> tmRndLng = new TreeMapTst<Long, Long>();
-        List<Long> rList = new ArrayList<Long>();
-        Long rnd;
-        while (tmRndLng.size() < 10000000) {
-            rnd = Math.round(Math.random() * 10000000L);
-            if (tmRndLng.put(rnd, rnd) == null) rList.add(rnd);
-        }
-        Calendar cEnd = Calendar.getInstance();
-        System.out.println("time 1.1 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        //
-        System.out.println("TreeMap");
-        cBegin = Calendar.getInstance();
-        TreeMapTst<Long, Long> tmLng = new TreeMapTst<Long, Long>();
-        for (int i = 0; i < 10000000; i++) {
-            rnd = rList.get(i);
-            tmLng.put(rnd, rnd);
-        }
-        System.out.println("maxLevel = " + tmLng.maxLevel);
-        System.out.println("RotateCount = " + tmLng.rotateCount);
-        cEnd = Calendar.getInstance();
-        System.out.println("time 2.1 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        //
-        System.out.println("RndBinTree");
-        cBegin = Calendar.getInstance();
-        RndBinTree<Long> avlLng = new RndBinTree<Long>();
-        for (int i = 0; i < 10000000; i++) {
-            rnd = rList.get(i);
-            avlLng.add(rnd);
-        }
-        System.out.println("maxLevel = " + avlLng.getMaxLevel());
-        System.out.println("RotateCount = " + avlLng.getRotateCount());
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.1 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        //
-        int i;
-        Iterator<Long> it;
-        //
-        /*
-        cBegin = Calendar.getInstance();
-        Iterator<Long> it = btLng.iterator();
-        i = 0;
-        while (it.hasNext()) {
-            it.next();
-            i++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.1 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + i + ") " + btLng.size());
-        */
-        //
-        cBegin = Calendar.getInstance();
-        it = tmLng.descendingKeySet().iterator();
-        i = 0;
-        while (it.hasNext()) {
-            it.next();
-            i++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.2 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + i + ") " + tmLng.size());
-        //
-        cBegin = Calendar.getInstance();
-        it = avlLng.iterator();
-        i = 0;
-        while (it.hasNext()) {
-            it.next();
-            i++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.3 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + i + ") " + avlLng.getSize());
-    }
-
-    public void testBinTree8() {
-        Calendar cBegin = Calendar.getInstance();
-        Long rnd;
-        List<Long> rList = new ArrayList<Long>();
-        for (int i = 0; i < 10000000; i++) {
-            rnd = Math.round(Math.random() * 10000000L);
-            rList.add(rnd);
-        }
-        Calendar cEnd = Calendar.getInstance();
-        System.out.println("time 1 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        //
-        cBegin = Calendar.getInstance();
-        BinTreeW<Long> btLng = new BinTreeW<Long>();
-        for (int i = 0; i < 10000000; i++) {
-            rnd = rList.get(i);
-            btLng.add(rnd);
-        }
-        System.out.println("maxLevel = " + btLng.getMaxLevel());
-        cEnd = Calendar.getInstance();
-        System.out.println("time 2.1 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        //
-        cBegin = Calendar.getInstance();
-        TreeSetTst<Long> tmLng = new TreeSetTst<Long>();
-        for (int i = 0; i < 10000000; i++) {
-            rnd = rList.get(i);
-            tmLng.add(rnd);
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 2.2 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        //
-        int i;
-        //
-        cBegin = Calendar.getInstance();
-        Iterator<Long> it = btLng.iterator();
-        i = 0;
-        while (it.hasNext()) {
-            it.next();
-            i++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.1 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + i + ") " + btLng.getSize());
-        //
-        cBegin = Calendar.getInstance();
-        it = tmLng.iterator();
-        i = 0;
-        while (it.hasNext()) {
-            it.next();
-            i++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.2 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + i + ") " + tmLng.size());
-    }
-
-    public void testBinTree9() {
-        Calendar cBegin = Calendar.getInstance();
-        Long rnd;
-        TreeSetTst<Long> rList1 = new TreeSetTst<Long>();
-        TreeSetTst<Long> rList2 = new TreeSetTst<Long>();
-        for (int i = 0; i < 15000000; i++) {
-            rnd = Math.round(Math.random() * 15000000L);
-            rList1.add(rnd);
-            rList2.add(rnd);
-        }
-        Calendar cEnd = Calendar.getInstance();
-        System.out.println("time 1 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        //
-        cBegin = Calendar.getInstance();
-        BinTreeW<Long> btLng = new BinTreeW<Long>();
-        Iterator<Long> itl = rList1.iterator();
-        while (itl.hasNext()) {
-            btLng.add(itl.next());
-        }
-        System.out.println("maxLevel = " + btLng.getMaxLevel() + "; rotateCount = " + btLng.getRotateCount());
-        cEnd = Calendar.getInstance();
-        System.out.println("time 2.1 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        //
-        cBegin = Calendar.getInstance();
-        TreeSetTst<Long> tmLng = new TreeSetTst<Long>();
-        Iterator<Long> it2 = rList2.iterator();
-        while (it2.hasNext()) {
-            tmLng.add(it2.next());
-        }
-        System.out.println("maxLevel = " + tmLng.getMaxLevel() + "; rotateCount = " + tmLng.getRotateCount());
-        cEnd = Calendar.getInstance();
-        System.out.println("time 2.2 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()));
-        //
-        int i;
-        //
-        cBegin = Calendar.getInstance();
-        Iterator<Long> it = btLng.iterator();
-        i = 0;
-        while (it.hasNext()) {
-            it.next();
-            i++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.1 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + i + ") " + btLng.getSize());
-        //
-        cBegin = Calendar.getInstance();
-        it = tmLng.iterator();
-        i = 0;
-        while (it.hasNext()) {
-            it.next();
-            i++;
-        }
-        cEnd = Calendar.getInstance();
-        System.out.println("time 3.2 = " + (cEnd.getTimeInMillis() - cBegin.getTimeInMillis()) + " (" + i + ") " + tmLng.size());
-    }
-
     public void testBinTreeEmpty() {
         BinTreeW<Long> btLng = new BinTreeW<Long>();
         Iterator<Long> it1 = btLng.iterator();
@@ -1719,45 +1237,6 @@ public class SimpleTest extends TestCase {
 
             }
         });
-    }
-
-    public static class BinTreeCheckPassEventTest<T extends Comparable<T>> implements BinTreeCheckPassEvent<T> {
-
-        private String errorMessage;
-        private boolean ignor = false;
-
-        public BinTreeCheckPassEventTest() {
-        }
-
-        public BinTreeCheckPassEventTest(boolean ignor) {
-            this.ignor = ignor;
-        }
-
-        @Override
-        public void onPass(
-                BinTreeIterator<T> leftIterator,
-                BinTreeIterator<T> rightIterator,
-                BinTreeNodeInterface<T> currentNode,
-                BinTreeNodeInterface<T> previousNode) {
-            if (!ignor)
-                System.out.println(errorMessage
-                        + "\r\n Current node: " + currentNode
-                        + "\r\n Left node: " + currentNode.getLeft()
-                        + "\r\n Right node: " + currentNode.getRight()
-                        + "\r\n Previous node: " + previousNode
-                        + "\r\n Left iterator: " + leftIterator
-                        + "\r\n Right iterator: " + rightIterator);
-        }
-
-        @Override
-        public void setErrorMessage(String errMsg) {
-            errorMessage = errMsg;
-        }
-
-        @Override
-        public String getErrorMessage() {
-            return errorMessage;
-        }
     }
 
 }
